+++ package.json
@@ -17,7 +17,8 @@
 		"mongodb": "^6.1.0",
 		"node-fetch": "^3.3.2",
 		"react": "^18.2.0",
-		"react-dom": "^18.2.0"
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.17.0"
 	},
 	"devDependencies": {
 		"@types/react": "^18.2.15",
