+++ src/RegionalForecast.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import BuoyData from './BuoyData';
+
+function RegionalForecast() {
+  return (
+    <div>RegionalForecast
+      <BuoyData currentStationId="46253" />
+    </div>
+  );
+}
+
+export default RegionalForecast;
